+++ packages/common/src/initiatives/_internal/InitiativeSchema.ts
@@ -1,12 +1,19 @@
+import { MetricSchema } from "../../core/schemas/internal/metrics/MetricSchema";
 import { IConfigurationSchema } from "../../core";
 import { HubItemEntitySchema } from "../../core/schemas/shared/HubItemEntitySchema";
 import { HubEntityStatus } from "../../types";
+import {
+  IF_SOURCE_TITLE_THEN_SOURCE_LINK,
+  IF_STATIC_THEN_REQUIRE_VALUE,
+  VALUE_TYPE_MAPPING,
+} from "../../core/schemas/internal/metrics/definitions";
 
 export type InitiativeEditorType = (typeof InitiativeEditorTypes)[number];
 export const InitiativeEditorTypes = [
   "hub:initiative:edit",
   "hub:initiative:create",
   "hub:initiative:associations",
+  "hub:initiative:metrics",
 ] as const;
 
 /**
@@ -25,5 +32,15 @@ export const InitiativeSchema: IConfigurationSchema = {
       default: HubEntityStatus.notStarted,
       enum: Object.keys(HubEntityStatus),
     },
+    _metric: {
+      type: "object",
+      required: ["cardTitle"],
+      properties: MetricSchema.properties,
+    },
   },
+  allOf: [
+    IF_SOURCE_TITLE_THEN_SOURCE_LINK,
+    IF_STATIC_THEN_REQUIRE_VALUE,
+    VALUE_TYPE_MAPPING,
+  ],
 } as IConfigurationSchema;