+++ packages/common/src/search/_internal/hubSearchItemsHelpers/getOgcCollectionUrl.ts
@@ -15,6 +15,11 @@ import { IApiDefinition } from "../../types/types";
  */
 export function getOgcCollectionUrl(query: IQuery, options: IHubSearchOptions) {
   const apiDefinition = options.api as IApiDefinition;
+  // Discussion posts as a target entity will be searchable with one collection,
+  // so simply use that for the URL
+  if (query.targetEntity === "discussionPost") {
+    return `${apiDefinition.url}/collections/discussion-post`;
+  }
   const collectionId = query.collection || "all";
   return `${apiDefinition.url}/collections/${collectionId}`;
 }