+++ packages/common/CHANGELOG.md
@@ -1,3 +1,39 @@
+# @esri/hub-common [14.103.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.102.0...@esri/hub-common@14.103.0) (2024-03-11)
+
+
+### Features
+
+* scaffold initiative metrics ([#1429](https://github.com/Esri/hub.js/issues/1429)) ([891f80a](https://github.com/Esri/hub.js/commit/891f80a5d0cbc54976ec1ae665e3315b362c5c03))
+
+# @esri/hub-common [14.102.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.101.0...@esri/hub-common@14.102.0) (2024-03-08)
+
+
+### Features
+
+* add orgId to IHubGroup ([#1434](https://github.com/Esri/hub.js/issues/1434)) ([66c574d](https://github.com/Esri/hub.js/commit/66c574d3edac0a0ebe0a1600fe3f8f8e7dc980c5))
+
+# @esri/hub-common [14.101.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.100.1...@esri/hub-common@14.101.0) (2024-03-07)
+
+
+### Features
+
+* **hub-common:** update events types ([#1426](https://github.com/Esri/hub.js/issues/1426)) ([83beb44](https://github.com/Esri/hub.js/commit/83beb44affeac6bced315680d41db63dd512aec9))
+
+## @esri/hub-common [14.100.1](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.100.0...@esri/hub-common@14.100.1) (2024-03-07)
+
+
+### Bug Fixes
+
+* remove unused scope on subblock section in stat card editor uiSchema ([#1431](https://github.com/Esri/hub.js/issues/1431)) ([9389cd8](https://github.com/Esri/hub.js/commit/9389cd8c9dfe4c8fdc6231c4ff227f7ff8f062b5))
+
+# @esri/hub-common [14.100.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.99.0...@esri/hub-common@14.100.0) (2024-03-07)
+
+
+### Features
+
+* **hub-common:** introduces types and function modifications for discussion post search ([#1425](https://github.com/Esri/hub.js/issues/1425)) ([b0080ff](https://github.com/Esri/hub.js/commit/b0080ff67de8526b40e91e8dd66aeb1074369f36))
+* **hub-common:** update feedback util ([#1433](https://github.com/Esri/hub.js/issues/1433)) ([f71a855](https://github.com/Esri/hub.js/commit/f71a8558e6bc37daccff6e9c38e81a50c6361382))
+
 # @esri/hub-common [14.99.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@14.98.1...@esri/hub-common@14.99.0) (2024-03-06)